Move wallet drawer class helper to utils
